+++ bacula/src/dird/verify.c
@@ -803,7 +803,7 @@ void get_attributes_and_compare_to_catalog(JCR *jcr, JobId_t JobId)
           * When ever we get a digest it MUST have been
           * preceded by an attributes record, which sets attr_file_index
           */
-         if (jcr->FileIndex != (uint32_t)file_index) {
+         if (jcr->FileIndex != file_index) {
             Jmsg2(jcr, M_FATAL, 0, _("MD5/SHA1 index %d not same as attributes %d\n"),
                file_index, jcr->FileIndex);
             free_pool_memory(fname);
